+++ stream_lib/include/IoStream.hpp
@@ -282,7 +282,7 @@ public:
      * For an output stream, frames available are the number of empty frames available
      * for the application to write.
      */
-    virtual android::status_t getFramesAvailable(uint32_t &avail,
+    virtual android::status_t getFramesAvailable(size_t &avail,
                                                  struct timespec &tStamp) const = 0;
 
     IStreamRoute *getCurrentStreamRoute() const { return mCurrentStreamRoute; }
